refactor(native): Python - extract calling
