+++ packages/cubejs-backend-native/src/python/mod.rs
@@ -4,5 +4,7 @@ mod entry;
 pub(crate) mod linux_dylib;
 pub(crate) mod python_model;
 pub(crate) mod runtime;
+pub mod utils;
 
 pub use entry::python_register_module;
+pub use utils::{python_fn_call_sync, python_obj_call_sync, python_obj_method_call_sync};
